finish backout api
Big commit, should have done this yeasterday but stuff it, I had a birthday to
celebrate.

Anyway, This app now uploads to lambda, that was a pain to configure.
The risk type api is now done in full, I hope.
Added token authentication

More details can be found in the documentation.

To the bar!